requirements.txt
Flask==2.2.5
flask-login==0.6.3
flask-mail==0.9.1
flask-mongoengine==1.0.0
mongoengine==0.27.0
python-dotenv==1.0.0
Werkzeug==2.3.8
itsdangerous==2.1.2
pymongo==4.6.1
dnspython==2.4.2
email-validator==2.1.0.post1
cloudinary==1.36.0
gunicorn==21.2.0